optimize findWhatFuckingSectionThisIs
This function looks like it's called pretty often, so optimizations here are pretty important. The main bottleneck I'm guessing is the showSearchHistory check. However, that's not saying the rest of this function doesn't have an impact. The old if statement would have originally done a lot of checks multiple times. My new version should retain the same behavior and check less. Also; viewForSupplementaryElementOfKind runs this twice, I don't think it needs to.